fix error in type deduction from onnx
Should now no longer require specifying the input type if onnx is a f16
model.

+++ compiler/src/denox/compiler/frontend/onnx/details/values/Dtype.cpp
@@ -69,7 +69,7 @@ std::optional<TensorDataType> dtype::details::Dtype::toTensorType() const {
   case DtypeKind::Float32:
     return TensorDataType::Float32;
   case DtypeKind::Float16:
-    return TensorDataType::Float32;
+    return TensorDataType::Float16;
   case DtypeKind::Undefined:
   case DtypeKind::Int8:
   case DtypeKind::Int16:
